♻️ Refactor: api 함수 위치 변경
refactor: api 함수 위치 변경
