+++ src/pages/QuestionForm/NPSPage.tsx
@@ -10,7 +10,7 @@ import {
 import { useState } from "react";
 import { useNavigate } from "react-router-dom";
 import { useSurvey } from "../../contexts/SurveyContext";
-import { createSurveyQuestion } from "../../service/form";
+import { createSurveyQuestion } from "./api";
 import { useQuestionByType } from "./hooks/useQuestionByType";
 
 export const NPSPage = () => {
